[MONDRIAN-2208] - Impala / Hive Dialect doesn't quote TIMESTAMP values properly (5.2 Suite)
Neither Hive nor Impala support ANSI timestamp literal.
The following construct should be used instead:
	cast( 'YYYY-MM-DD hh:mm:ss[fffff]' as timestamp )

This is what was implemented in HiveDialect.
The unit test is too strightforward.
The same case is covered in mondrian-tck NativeFilterTest.testCompoundPredicateNoJoinsDateLiteralSyntax().
The problem with tck is that Foodmart schema for hive does not contain neither DATE nor TIMESTAMP columns.

That should be taken into account that Hive2 supports DATE datatype and ANSI literal for it.
In order to leverage this a separate Dialect implementation should be done (how to distiguish between Hive1 and Hive2?)

The fix was manually verified with Analyzer for Hive1 (0.8), Hive2 (0.12) and Impala on Cludera 4.7 and 5.1