[MONDRIAN-2376] distinct-count measures could cause an agg table miss if levels involved are set with uniqueMembers=true
First noticed when a cube was set with a distinct-count default
measure, which caused unexpected agg table misses during
SqlTupleReader queries (which in many cases will be evaluated in the
context of the default measure).

This occurred because the levelBitKey would be set with just the child
level's bit set, even though the aggregate tables bitkey would include
each parent level as well.  It's okay to match when the parent bitkeys
are set in this case, because presence of the parent columns does not
effect granularity.